requirements.txt
discord.py>=1.7.1
python-decouple==3.4
pillow
pyrebase
imageio
akinator.py
pygifsicle
fun-responses
discord-components
asyncio
anime-images-api
disputils
pet-pet-gif
youtube_dl
pynacl
ffmpeg
DiscordUtils
weeby.py
